Removing extended attrs from short OCCI rendering
Extended model attributes such as `title` or `location` should not be included when using short rendering of categories (e.g., in instances). These are present only in the full category rendering (model available at `/-/`). This change should not have any impact on the current generation of OCCI servers, it is intended to maintain compatibility with the new generation, with stricter parsing rules.